全面替换工具
