+++ docs/.vitepress/theme/data/doc.data.ts
@@ -53,50 +53,6 @@ export default createContentLoader("**/*.md", {
                     : (raw.frontmatter.createdAt instanceof Date && !isNaN(raw.frontmatter.createdAt.getTime()))
                         ? raw.frontmatter.createdAt.getTime()
                         : undefined;
-                
-                // Filter resources
-                const rawResources: Record<string, ResourceInput> = 
-                    (typeof raw.frontmatter.resources === 'object' && raw.frontmatter.resources !== null)
-                        ? raw.frontmatter.resources
-                        : {};
-                const resources: Record<string, ResourceData> = {};
-
-                Object.entries(rawResources).forEach(([key, value]) => {
-                    if (value === false) {
-                        resources[key] = { url: false };
-                    } else if (typeof value === 'string') {
-                        resources[key] = {
-                            url: value,
-                            label: value
-                        }
-                    } else if (typeof value === 'object') {
-                        resources[key] = {
-                            url: (typeof value.url === 'string' || value.url === false)
-                                ? value.url
-                                : undefined,
-                            label: (typeof value.label === 'string')
-                                ? value.label
-                                : undefined,
-                            icon: (
-                                typeof value.icon === 'string' ||
-                                (
-                                    typeof value.icon === 'object' &&
-                                    value.icon !== null &&
-                                    typeof value.icon.component === 'string'
-                                )
-                            )
-                                ? value.icon
-                                : undefined,
-                            download: (typeof value.download === 'boolean' || typeof value.download === 'string')
-                                ? value.download
-                                : undefined,
-                            order: ("order" in value) ? any2Number(value.order) : undefined,
-                            type: (["file", "image", "website", "download"].includes(value.type || ""))
-                                ? value.type
-                                : undefined,
-                        }
-                    }
-                })
 
                 return {
                     ...raw,
@@ -113,7 +69,7 @@ export default createContentLoader("**/*.md", {
                     inherit: (raw.frontmatter.inherit === undefined)
                         ? undefined
                         : !!raw.frontmatter.inherit,
-                    resources,
+                    resourcesList: [raw.frontmatter.resources],
                     treeTitle: (typeof raw.frontmatter.treeTitle === 'string')
                         ? raw.frontmatter.treeTitle
                         : undefined,
